backend_backup_20251004_140040/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic-settings==2.0.3
motor==3.3.2
pymongo==4.6.0
redis==5.0.1
rq==1.15.1
requests==2.32.5
numpy==1.24.3
soundfile==0.12.1
pydub==0.25.1
loguru==0.7.2
python-dotenv==1.0.0
orjson==3.9.10
slowapi==0.1.9
fastapi-cors==0.0.6
google-cloud-storage==3.3.1
beanie==1.25.0
pytz==2023.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0
